Tests/LibWeb/Text/expected/css/shadow-slotted-animation-name-shared-constructable-sheet-prefers-document-keyframes.txt

opacity: 0.5
